requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
google-auth==2.26.0
boto3==1.34.0
botocore==1.34.0
youtube-transcript-api>=0.6.3
python-dotenv>=1.0.0
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
